+++ assets/javascripts/kitten/components/form/text-input-with-unit-form.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react'
+import React from 'react'
 import Radium, { StyleRoot } from 'radium'
 import PropTypes from 'prop-types'
 import { Marger as MargerBase } from 'kitten/components/layout/marger'
@@ -16,7 +16,6 @@ const Marger = Radium(MargerBase)
 export const TextInputWithUnitForm = ({
   align,
   formIsDisabled,
-  viewportIsSOrLess,
   inputId,
   inputPlaceholder,
   inputUnit,
@@ -27,24 +26,30 @@ export const TextInputWithUnitForm = ({
   buttonLabel,
   onButtonMouseEnter,
   onButtonMouseLeave,
-  onButtonClick,
   onFormSubmit,
   onInputBlur,
   onInputChange,
   onInputFocus,
+  version,
 }) => {
   const formStyle = align === 'center' && styles.form.centered
+  const isTinyVersion = version === 'tiny'
+
+  const gridColProps = isTinyVersion ? {} : { 'col-xs': 7, 'col-m': 5 }
+
+  const buttonStyles = isTinyVersion ? styles.button.tinyVersion : styles.button
 
   return (
-    <form onSubmit={onFormSubmit}>
+    <form onSubmit={onFormSubmit} style={styles.form.grid}>
       <Marger top="3" bottom={!inputIsOnError ? 3 : 1}>
         <Grid style={formStyle}>
-          <GridCol col-xs="7" col-m="5">
+          <GridCol {...gridColProps}>
             <Marger bottom="1.5">
               <Label size="micro" htmlFor={inputId}>
                 {inputLabel}
               </Label>
             </Marger>
+
             <Marger top="1.5" bottom={inputIsOnError ? 1 : null}>
               <TextInputWithUnit
                 error={inputIsOnError}
@@ -60,6 +65,7 @@ export const TextInputWithUnitForm = ({
                 autoComplete="off"
               />
             </Marger>
+
             {inputIsOnError && (
               <Marger top="1">
                 <Text size="micro" color="error" weight="regular">
@@ -70,6 +76,7 @@ export const TextInputWithUnitForm = ({
           </GridCol>
         </Grid>
       </Marger>
+
       <Marger style={formStyle}>
         <StyleRoot>
           <Button
@@ -77,7 +84,7 @@ export const TextInputWithUnitForm = ({
             modifier="helium"
             type="submit"
             aria-label={buttonLabel}
-            style={styles.button}
+            style={buttonStyles}
             onMouseEnter={onButtonMouseEnter}
             onMouseLeave={onButtonMouseLeave}
             disabled={formIsDisabled}
@@ -109,6 +116,8 @@ TextInputWithUnitForm.propTypes = {
   align: PropTypes.string,
   formIsDisabled: PropTypes.bool,
   onFormSubmit: PropTypes.func,
+
+  version: PropTypes.oneOf(['default', 'tiny']),
 }
 
 TextInputWithUnitForm.defaultProps = {
@@ -124,10 +133,16 @@ TextInputWithUnitForm.defaultProps = {
   formIsDisabled: false,
   onFormSubmit: () => {},
   align: 'center',
+  version: 'default',
 }
 
 const styles = {
   form: {
+    grid: {
+      margin: 0,
+      padding: 0,
+    },
+
     centered: {
       display: 'flex',
       justifyContent: 'center',
@@ -136,6 +151,10 @@ const styles = {
 
   button: {
     [`@media (max-width: ${ScreenConfig['XS'].max}px)`]: {
+      width: '100%',
+    },
+
+    tinyVersion: {
       width: '100%',
     },
   },